requirements.txt
# Core dependencies
ipython
jupyterlab
matplotlib
mkdocs
notebook
numpy
pandas
pip
python-dotenv
ruff
scikit-learn

# Geospatial processing
geopandas
shapely

# XML processing
lxml

# Configuration & validation
pydantic>=2.0
PyYAML

# Data formats
pyarrow

# Logging
loguru
tqdm

# Install package in editable mode
-e .
